Image grid, and JS for hamburger menu button
Started to design an image grid using CSS. Began to incorporate JS for the hamburger menu (edited html & css code for this too); position property,  added in extra classes to style hamburger menu. JS: function for hamburer menu button, query selectors to display menu when button is clicked and hide it when not.